[#7] Move mouse cursor to position.
This commit adds the ability to move the mouse cursor to a given position
and changes the interaction with the event queue slightly. Adding an event
is done by putting a vector like [:click] or [:move-to 100 200] on the
publicly available channel in the io namespace. Whether this is how it's
going to stay is an open question, but it works well for now.