[#11] Compute color histogram of image.
It is now possible to compute a simple color histogram of a given
BufferedImage. The histogram uses 8 buckets for each color, and each
bucket contains a normalized number between 0 and 255. The 8 buckets
sum up to 255. Why 255? It will make it really easy to create reader
tags with hexadecimal literals later on, which can be entered in the
EDN files without too much verbosity. I envision a reader tag like

`#cljck.color.Histogram[0x0011223344556677 0x8899aabbccddeeff
                        0x0123456789abcdef]`

which is a fairly compact notation for a histogram.